Add files via upload
create the folder for the files in the location where the index page is placed. Create folder with name "css" and place the . css files create the folder with name "js" and place the js file in it. create folder "images" and place the images and logo in it. create the folder "links" and place the php and link files. create folder "participant_images" and leave it as empty for storing the registering participant image.